+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- An uncompressed LZMA2 chunk is now handed over as it is copied, instead of only once the dictionary fills.
+
 ## 0.20.0 - 2026-08-23
 
 ### Added